Added the dispatch code for mutex.
As mentioned in the comment, a dynamic mechanism to handle IDs and
allow use of names must be added ( someday ). Unless we switch to
XPCOM or a similar object model first.

At this moment I don't have any urge to compile IDLs, and the
overhead seems too big, we can still keep it simple.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297344 13f79535-47bb-0310-9956-ffa450edef68