- removed the mutex code, the normal invocation mechanism will be used ( i.e. the ajp 'bridge' ).
- cleanup a bit the unix socket, preparing for the removal ( again the same
mechanism will be used - I don't want to do the direct buffer and optimizations
in 2 places )

- added some of the methods that were missing. Note that signal is not complete,
I want ( again ) to use the same code for java-C invocation.

Based on the various tests I made so far and what I've seen in other systems,
'straight' use of JNI is not a good solution ( look at Mozilla and all other
projects that integrate java with C applications ). There are some fine
optimizations that can be made in JDK1.4, and some other that would
work in 1.3 - but it's easier to code the critical code in few methods
( the 2 methods that call java from C and C from java ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297356 13f79535-47bb-0310-9956-ffa450edef68