A bunch of changes - mostly to get around the shm behavior on HPUX.
On some HPUX-es ( and possibly other platforms ) the mmap will fail if the
child is mapping the same file. The solution is to map in the parent and
inherit it in the child. That's what we do - using the top apr pool
to store a note with the shm pointer.

I also added a command to jkstatus to allow reseting the endpoint counters
( for tunning you typically run few 1000s requests to 'warm up' the server,
and then do the real tests - the jkstatus gives a very good info on how
much time is spent on server - and how much is in jk versus tomcat )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297403 13f79535-47bb-0310-9956-ffa450edef68