Added a flag to select the byte[] access mechanism.
getArrayElements is the fastest - if pinning is supported by the VM. Aparently
JDK1.4 doesn't pin, so the performance will go down.

getArrayRegion works well on all VMs, so it is the default.

The third option will work only on JDK1.4, with direct buffers ( it'll probably have
the same performance with getArrayElements with pinning ). For this
to work we need to implemnet a direct buffer version for ByteChunk and
MsgAjp.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297415 13f79535-47bb-0310-9956-ffa450edef68