A bit of refactoring.
All workers2.properties parsing/saving code ( i.e. the properties + [ini] extensions )
are now in jk_config_file.

jk_config.c contains all generic methods that 'emulate' the bean ( or ant)-like
configuration of components.

Any alternate config repository can be used - the requirement is to be
able to create a map with component names as keys and an attribute map
as value.

( that's not actually required - you can walk a registry and call processNode
if you want ).

Note that jk_map.c is back to 'map only', and may be deprecated ( since
apr has better maps ). We keep using it for now only because it's stable
and doesn't create problems ( and we can store void * and char *)


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297425 13f79535-47bb-0310-9956-ffa450edef68