Initial code for the (experimental for now ) JMX support for jk.
This is the first step in enabling JMX support for the C components
and the remote native jk ( there is still a _lot_ of work for this to
work ). However all java side jk components should magically become
JMX aware.

Of course this is not enabled by default nor finalized ( there are lots
of debug infos ). Just some code that may sit in the jar, unused.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297458 13f79535-47bb-0310-9956-ffa450edef68