Another attempt to fix the 'tomcat restart' problem.
send ( with either write or send() ) doesn't detect if tomcat is disconnected -
only the first receive does report the error.

What we do is use the 'recoverable' field to mark if tomcat has already
started to process the request ( i.e. we received at least one packet -
probably the HEAD, in which case the error can't be recovered ).

If this is just a tomcat restart, we'll save the post body ( initial chunk ),
and resend it. The logic is a bit tricky - but I think I got it right this
time ( it affected the case that a POST is made just after restart )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297579 13f79535-47bb-0310-9956-ffa450edef68