Tentative fix for 12346.
If an unrecoverable error happens ( for example when the client
hits stop - the server can't send more data since the connection is lost )
we need to forcefully break the ajp13 connection, since tomcat can't
know this and will continue to send data.

This behavior is a result of the optimizations made for the 'normal' case,
i.e. the fact that tomcat doesn't wait for confirmation when sending
chunks of data. Adding the roundtrip will have big negative performance hit -
and it's better to deal with the error cases.

Note that the alternative ( and what seems to happen for apache ) is to
ignore the server errors and continue to receive chunks and ignore them.
That can save the ajp connection - with the price of having tomcat send
useless data. For a large file that may be a bad solution, and tomcat will
not be notified that the client had problems ( which may be a usefull info ).

If this doesn't fix the problem - please send me logs with ajp debug
enabled. I can't reproduce it ( or run IIS ), but this is clearly a bug..


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@297995 13f79535-47bb-0310-9956-ffa450edef68