First cleanup/refactoring.
Instead of | separated values ( I can't believe I did that ! ) use the same
format as the config. This way I may get around and implement it for config
too ( part of modeler probably ).

Added initial support for methods - init/destroy is all I care for jk, but
it seems like this may become a very nice piece of reusable code to manage
multiple instances.

One thing that may be very interesting is to do a set on startup and
shutdown ( disabled ? ).

Testing mod_jk from the JMX console showed a lot of small problems
in config - mostly ordering. The "classic" config is still the best,
but this will be really cool.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@298433 13f79535-47bb-0310-9956-ffa450edef68