Adding attribute "distance" to balanced workers.
Express Preferences between the balanced workers of an lb worker.
A load balancer will never choose some balanced worker
in case there is another usable worker with lower distance.

Only in case all workers below a given distance are in error, disabled or stopped,
workers of a larger distance are eligible for balancing.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@407096 13f79535-47bb-0310-9956-ffa450edef68