Refactor "route" return value from LB to httpd note JK_WORKER_ROUTE.
1) Our code always returned the worker route, except when the worker
was chosen via the domain. Then it returns the domain name.
2) We used a member of the worker_record, as an intermediate store,
but the worker_record is not thread safe, only the service record is.

So we only need to return the information, if the balancer used the
domain to find the worker and set the service route to the domain
in this case, otherwise to the worker route.
Use an int pointer argument to return the info and remove the
worker_record member "r".


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@610561 13f79535-47bb-0310-9956-ffa450edef68