Force copy of automake files when running buildconf.sh.
Some of the helper files come with libtool
and with automake. We installed them forced
from libtool, and then the missing ones from automake.

Now let automake overwrite the files it has itself,
because automake is the source for these files,
so it's likely the ones installed with automake
are newer than the libtool ones.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@774597 13f79535-47bb-0310-9956-ffa450edef68